pyproject.toml
[project]
name = "content-studio-agent"
version = "0.1.0"
description = "Multi-agent social media content creation platform powered by Google ADK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.13.0",
    "google-genai>=1.17.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.9",
    "pillow>=10.0.0",
    "colorthief>=0.2.1",
    "jinja2>=3.1.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "tools", "memory"]
